transforms: add transducers to operate on collections of ByteBufs
net exposes data as channels of ByteBuf instances in several places.
For instance, by default the body of Http requests will be exposed
in this manner.

When large request bodies are sent, there might be the need for
transforms to happen on the body without accumulating all ByteBuf
instances.

In this case, transducers which handle ByteBuf reference counting
are necessary.

This patch provides the foundation to build these transforms by
creating correctly counted CompositeByteBuf instances.