channel: support bridging netty Channels to core.async directly
This is a first step to reduce the amount of glue necessary between
core.async and netty. Instead of pushing incoming messages down
a channel, we listen for take! events and schedule reads accordingly
on the Netty channel.

A concurrent queue allows messages to be buffered as well as pending
takes..